% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQuality.R
\name{checkDataQuality}
\alias{checkDataQuality}
\title{Check data quality}
\usage{
checkDataQuality(strToPrint)
}
\arguments{
\item{strToPrint}{Some string to print out with the}
}
\value{
Nothing; just prints something on-screen
}
\description{
This function checks data quality
}
