% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQuality.R
\name{treatMissingValues}
\alias{treatMissingValues}
\title{Treat missing values}
\usage{
treatMissingValues()
}
\value{
Nothing; just prints something on-screen
}
\description{
This function simulates missing values treatment
}
